WEBNEW-171 :art: changed code as per peer review, apart from url
