+++ src/blog/article/BlogFooter.tsx
@@ -1,19 +1,7 @@
 import React from 'react'
-import styled from 'styled-components'
-import theme from '../../theme/theme'
-import { Input } from '../../components/input'
 import Author from './Author'
 import { Summary, BlogFooterContainer } from './BlogFooter.styles'
-import { PageContentProps } from './PageContent'
-
-export const StyledInput = styled(Input)`
-  label {
-    color: ${theme.colors.darkGrey};
-  }
-  input {
-    color: ${theme.colors.darkGrey};
-  }
-`
+import { PageContentProps } from './PageContent.types'
 
 export const BlogFooter: React.FC<Pick<PageContentProps, 'author'>> = ({
   author,
