+++ src/blog/article/PageContent.styles.tsx
@@ -1,17 +1,8 @@
 import styled from 'styled-components'
 import { ShareBar } from '../../components/shareBar'
-import { Column } from '../../components/grid'
 import { space } from '../../theme/space'
 import { mediaQueries } from '../../theme/mediaQueries'
 
-export const Content = styled(Column)`
-  max-width: 830px;
-  margin: 0 auto;
-  margin-top: -140px;
-  background-color: white;
-  position: relative;
-  padding: 30px;
-`
 export const ShareBarWrapper = styled.div`
   display: none;
   position: absolute;
