+++ src/blog/article/PageContent.tsx
@@ -1,35 +1,10 @@
 import React from 'react'
-import { Document } from '@contentful/rich-text-types'
 import { Wrapper } from '../../components/wrapper/Wrapper'
 import { BlogHeader } from './BlogHeader'
 import { BlogContent } from './blogContent'
 import { BlogFooter } from './BlogFooter'
 import { ShareBarWrapper, DesktopShareBar } from './PageContent.styles'
-export interface Category {
-  title: string
-  hexColor: string
-}
-
-interface Thumbnail {
-  file: { url: string }
-  id: string
-}
-interface Author {
-  displayName?: string
-  thumbnail?: Thumbnail
-  jobTitle?: string
-}
-
-interface Article {
-  json: Document
-}
-export interface PageContentProps {
-  article: Article
-  title: string
-  datePublished: string
-  category: Category
-  author: Author
-}
+import { PageContentProps } from './PageContent.types'
 
 export const PageContent: React.FC<PageContentProps> = ({
   article,
