+++ src/blog/article/blogContent/BlogContent.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { documentToReactComponents } from '@contentful/rich-text-react-renderer'
-import { PageContentProps } from '../PageContent'
+import { PageContentProps } from '../PageContent.types'
 import renderMethods from './renderMethods'
 
 export const BlogContent: React.FC<Pick<PageContentProps, 'article'>> = ({
